backend/requirements.txt
asgiref==3.11.0
Django==5.2.8
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.3
django-cors-headers==4.9.0
requests==2.31.0
pandas==2.1.4
dj-database-url==2.1.0
psycopg2-binary==2.9.9